Extract inference API payload serialization into dedicated module.
Move summary formatting and solution wire shaping out of analytic.py to keep turn orchestration separate from response assembly.
