+++ packages/api/api/analytics/military_score_inference/inference_api_payload.py
@@ -1,0 +1,193 @@
+"""API payload serialization for military score build inference results."""
+
+from api.analytics.military_score_inference.actions import ActionCatalog
+from api.analytics.military_score_inference.models import (
+    InferenceObservation,
+    InferenceProblem,
+    InferenceResult,
+    InferenceSolution,
+)
+from api.analytics.military_score_inference.score_arithmetic import (
+    solution_military_score_arithmetic_payload,
+)
+from api.analytics.military_score_inference.solver import (
+    STATUS_INVALID_PROBLEM,
+    STATUS_NO_EXACT_SOLUTION,
+    STATUS_TIME_LIMITED,
+)
+from api.models.game import TurnInfo
+
+STATUS_NO_PRIOR_TURN = "no_prior_turn"
+STATUS_SOLVER_ERROR = "solver_error"
+
+
+def inference_result_to_api_payload(
+    result: InferenceResult,
+    catalog: ActionCatalog,
+    observation: InferenceObservation,
+    turn: TurnInfo,
+    problem: InferenceProblem,
+    *,
+    policy_steps_attempted: list[str] | None = None,
+    step_diagnostics: list[dict[str, object]] | None = None,
+    extra_diagnostics: dict[str, object] | None = None,
+) -> dict[str, object]:
+    """Shape a solver result into the Core scores row inference object."""
+    from api.analytics.military_score_inference.analytic import (
+        build_inference_solver_diagnostics,
+    )
+
+    solver_diagnostics = {
+        "status": result.status,
+        **result.diagnostics,
+    }
+    diagnostics = build_inference_solver_diagnostics(
+        turn=turn.settings.turn,
+        observation=observation,
+        problem=problem,
+        catalog=catalog,
+        turn_info=turn,
+        solver=solver_diagnostics,
+        extra={
+            "policy_steps_attempted": policy_steps_attempted or [catalog.policy_step_id],
+            "policy_step_attempts": step_diagnostics or [],
+            **(extra_diagnostics or {}),
+        },
+    )
+    return _inference_api_payload(
+        status=result.status,
+        summary=format_inference_summary(result),
+        solutions=result.solutions,
+        diagnostics=diagnostics,
+        observation=observation,
+        catalog=catalog,
+    )
+
+
+def format_inference_summary(result: InferenceResult) -> str:
+    """Return compact row-level summary text for the inference column."""
+    if result.status == STATUS_NO_PRIOR_TURN:
+        return "Prior turn score data unavailable"
+    if result.status == STATUS_INVALID_PROBLEM:
+        reason = result.diagnostics.get("reason")
+        if isinstance(reason, str) and reason:
+            return f"Invalid inference problem: {reason}"
+        return "Invalid inference problem"
+    if result.status == STATUS_NO_EXACT_SOLUTION:
+        return "No feasible build explanation found"
+    if result.status == STATUS_SOLVER_ERROR:
+        return "Build inference failed"
+    if result.status == STATUS_TIME_LIMITED and not result.solutions:
+        return "Inference timed out before finding a solution"
+    if not result.solutions:
+        return "No feasible build explanation found"
+
+    best_summary = _format_solution_brief(result.solutions[0])
+    alternative_count = len(result.solutions) - 1
+    if alternative_count == 0:
+        return f"Best: {best_summary}"
+    if alternative_count == 1:
+        return f"Best: {best_summary}; 1 alternative"
+    return f"Best: {best_summary}; {alternative_count} alternatives"
+
+
+def _format_solution_brief(solution: InferenceSolution) -> str:
+    parts: list[str] = []
+    for action in solution.actions:
+        if action.count == 1:
+            parts.append(action.label)
+        else:
+            parts.append(f"{action.count}x {action.label}")
+    for ship_build in solution.ship_builds:
+        if ship_build.count == 1:
+            parts.append(ship_build.label)
+        else:
+            parts.append(f"{ship_build.count}x {ship_build.label}")
+    return "; ".join(parts) if parts else "no actions"
+
+
+def _inference_api_payload(
+    *,
+    status: str,
+    summary: str,
+    solutions: tuple[InferenceSolution, ...],
+    diagnostics: dict[str, object],
+    observation: InferenceObservation | None = None,
+    catalog: ActionCatalog | None = None,
+) -> dict[str, object]:
+    return {
+        "status": status,
+        "summary": summary,
+        "solutionCount": len(solutions),
+        "isComplete": status != STATUS_TIME_LIMITED,
+        "solutions": (
+            [
+                _serialize_solution_with_arithmetic(observation, catalog, solution)
+                for solution in solutions
+            ]
+            if observation is not None and catalog is not None
+            else [_serialize_solution_without_arithmetic(solution) for solution in solutions]
+        ),
+        "diagnostics": diagnostics,
+    }
+
+
+def _serialize_solution_actions(
+    solution: InferenceSolution,
+) -> list[dict[str, object]]:
+    return [
+        {
+            "actionId": action.action_id,
+            "label": action.label,
+            "count": action.count,
+        }
+        for action in solution.actions
+    ]
+
+
+def _serialize_solution_ship_builds(
+    solution: InferenceSolution,
+) -> list[dict[str, object]]:
+    return [
+        {
+            "comboId": ship_build.combo_id,
+            "label": ship_build.label,
+            "count": ship_build.count,
+            "hullId": ship_build.hull_id,
+            "engineId": ship_build.engine_id,
+            "beamId": ship_build.beam_id,
+            "torpId": ship_build.torp_id,
+            "beamCount": ship_build.beam_count,
+            "launcherCount": ship_build.launcher_count,
+        }
+        for ship_build in solution.ship_builds
+    ]
+
+
+def _serialize_solution_core(solution: InferenceSolution) -> dict[str, object]:
+    return {
+        "objectiveValue": solution.objective_value,
+        "actions": _serialize_solution_actions(solution),
+        "shipBuilds": _serialize_solution_ship_builds(solution),
+    }
+
+
+def _serialize_solution_with_arithmetic(
+    observation: InferenceObservation,
+    catalog: ActionCatalog,
+    solution: InferenceSolution,
+) -> dict[str, object]:
+    actions_by_id = {action.id: action for action in catalog.aggregate_actions}
+    combos_by_id = {combo.combo_id: combo for combo in catalog.ship_build_combos}
+    payload = _serialize_solution_core(solution)
+    payload["militaryScoreArithmetic"] = solution_military_score_arithmetic_payload(
+        solution,
+        observation,
+        actions_by_id,
+        combos_by_id,
+    )
+    return payload
+
+
+def _serialize_solution_without_arithmetic(solution: InferenceSolution) -> dict[str, object]:
+    return _serialize_solution_core(solution)
